Move config access logic into their own package
