+++ src/main/java/org/cloudburstmc/server/Server.java
@@ -17,6 +17,9 @@ import lombok.val;
 import net.daporkchop.ldbjni.LevelDB;
 import org.cloudburstmc.server.command.CommandSender;
 import org.cloudburstmc.server.command.ConsoleCommandSender;
+import org.cloudburstmc.server.config.CloudburstYaml;
+import org.cloudburstmc.server.config.ServerConfig;
+import org.cloudburstmc.server.config.ServerProperties;
 import org.cloudburstmc.server.console.NukkitConsole;
 import org.cloudburstmc.server.entity.Attribute;
 import org.cloudburstmc.server.event.server.*;
